WT-2755 flexelint configuration treats size_t as 4B type (#2858)
KNF, remove space after "sizeof" keyword.

Info 790: Suspicious truncation, integral to float

strlen returns a size_t, cast before comparing against a wt_off_t.

size_t is 8B, more size_t cleanups.

+++ src/conn/conn_api.c
@@ -1554,7 +1554,7 @@ __conn_single(WT_SESSION_IMPL *session, const char *cfg[])
 		 */
 #define	WT_SINGLETHREAD_STRING	"WiredTiger lock file\n"
 		WT_ERR(__wt_filesize(session, conn->lock_fh, &size));
-		if (size != strlen(WT_SINGLETHREAD_STRING))
+		if ((size_t)size != strlen(WT_SINGLETHREAD_STRING))
 			WT_ERR(__wt_write(session, conn->lock_fh, (wt_off_t)0,
 			    strlen(WT_SINGLETHREAD_STRING),
 			    WT_SINGLETHREAD_STRING));
